Fix offset types.

+++ network_buffer.h
@@ -120,9 +120,9 @@ int
 simple_buffer_resize(struct simple_buffer * const buf, size_t newsize)
 {
 	if (newsize != buf->max_size) {
-		uint64_t head_offset = buf->head - buf->data;
-		uint64_t tail_offset = buf->tail - buf->data;
-		uint64_t userptr_offset = buf->userptr - buf->data;
+		size_t head_offset = buf->head - buf->data;
+		size_t tail_offset = buf->tail - buf->data;
+		size_t userptr_offset = buf->userptr - buf->data;
 		buf->max_size = (newsize/buf->chunk_size+1) * buf->chunk_size;
 		char *newbufdata = realloc(buf->data, buf->max_size);
 		if (newbufdata == NULL)
